// src/app/core/model/books.model.ts
export interface Book {
    id: string;
    volumeInfo: {
        title: string;
        authors: string[];
        publisher: string;
        publishedDate: string;
        description: string;
        pageCount: number;
        imageLinks: {
            thumbnail: string;
            smallThumbnail: string;
        }
    }
}